upload files to files/N/NNNNNNNN
where N is a number, and the folder is the same as the last N of the
timestamp.

if future, should either use sha-1 of timestamp and upload user, or
timestamp with fraction.

when fetching files (not supported yet), a counter should be used in
addition to the filename... or just use the id, just as it is now.

i have a link in delicious to apache/nginx protected file serving.